+++ paddle/fluid/pir/dialect/operator/interface/infer_symbolic_shape/binary_infer_sym.h
@@ -1,4 +1,4 @@
-// Copyright (c) 2024 PaddlePaddle Authors. All Rights Reserved.
+// Copyright (c) 2023 PaddlePaddle Authors. All Rights Reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -17,15 +17,17 @@
 #include "paddle/pir/include/dialect/shape/utils/shape_analysis.h"
 
 namespace paddle::dialect {
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Arange)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Data)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Empty)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Feed)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Full)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(FullIntArray)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Gaussian)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Randint)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(TrilIndices)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(TriuIndices)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Uniform)
+
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(Embedding)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(SparseWeightEmbedding)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(ExpandAs)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(Gather)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(GatherNd)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(Kron)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(MaskedSelect)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(Matmul)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(Searchsorted)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(TakeAlongAxis)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(TopPSampling)
+
 }  // namespace paddle::dialect
