--- paddle/fluid/pir/dialect/operator/interface/infer_symbolic_shape/paddle_op_infer_sym.cc
@@ -1,1022 +1,0 @@
-// Copyright (c) 2023 PaddlePaddle Authors. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#include "paddle/fluid/pir/dialect/operator/interface/infer_symbolic_shape/paddle_op_infer_sym.h"
-#include "paddle/common/ddim.h"
-#include "paddle/fluid/pir/dialect/operator/interface/infer_symbolic_shape/infer_sym_slice_utils.h"
-#include "paddle/fluid/pir/dialect/operator/interface/infer_symbolic_shape/infer_sym_utils.h"
-#include "paddle/fluid/pir/dialect/operator/ir/op_attribute.h"
-
-namespace paddle::dialect {
-
-bool ShapeOpInferSymbolicShape(pir::Operation *op,
-                               pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  const symbol::ShapeOrDataDimExprs &operand_shape_or_data =
-      shape_analysis->GetShapeOrDataForValue(op->operand_source(0));
-  const auto &out_data = operand_shape_or_data.shape();
-  const std::vector<symbol::DimExpr> shape{std::int64_t(out_data.size())};
-  symbol::ShapeOrDataDimExprs shape_or_data{
-      symbol::TensorShapeOrDataDimExprs(shape, out_data)};
-
-  shape_analysis->SetShapeOrDataForValue(op->result(0), shape_or_data);
-  return true;
-}
-
-bool ShapeSrOpInferSymbolicShape(
-    pir::Operation *op, pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  return ShapeOpInferSymbolicShape(op, shape_analysis);
-}
-
-bool StackOpInferSymbolicShape(pir::Operation *op,
-                               pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  pir::Value operand_source = op->operand_source(0);
-
-  const auto &attributes = op->attributes();
-  int axis = attributes.at("axis").dyn_cast<pir::Int32Attribute>().data();
-
-  const symbol::TensorListShapeOrDataDimExprs &shape_data_list =
-      shape_analysis->GetShapeOrDataForValue(operand_source)
-          .dyn_cast<symbol::TensorListShapeOrDataDimExprs>();
-
-  int rank = shape_data_list[0].shape().size();
-  if (axis < 0) axis += rank + 1;
-
-  const symbol::ShapeOrDataDimExprs shape_data = [&] {
-    std::vector<symbol::DimExpr> shape_dim_exprs;
-    std::vector<symbol::DimExpr> data_dim_exprs;
-    for (size_t i = 0; i < shape_data_list.size(); ++i) {
-      if (shape_data_list[i].data().has_value() && axis == 0) {
-        data_dim_exprs.emplace_back(shape_data_list[i].data().value()[0]);
-      }
-    }
-
-    if (!data_dim_exprs.empty()) {
-      shape_dim_exprs.emplace_back(
-          static_cast<std::int64_t>(shape_data_list.size()));
-    } else {
-      for (int i = 0; i < rank; ++i) {
-        details::BuildCstrEqForTensorListAlongAxis(
-            shape_analysis, shape_data_list, i);
-      }
-      shape_dim_exprs.insert(shape_dim_exprs.begin() + axis,
-                             static_cast<std::int64_t>(shape_data_list.size()));
-    }
-
-    return symbol::ShapeOrDataDimExprs(
-        symbol::TensorShapeOrDataDimExprs(shape_dim_exprs, data_dim_exprs));
-  }();
-
-  pir::Value res = op->result(0);
-  shape_analysis->SetShapeOrDataForValue(res, shape_data);
-  return true;
-}
-
-bool SumOpInferSymbolicShape(pir::Operation *op,
-                             pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  bool keepdim = GetBoolAttr(op, "keepdim");
-  bool reduce_all = false;
-
-  auto axis_gen_op = op->operand_source(1).defining_op();
-  if (axis_gen_op->isa<paddle::dialect::FullIntArrayOp>()) {
-    std::vector<int64_t> axis = details::GetVectorAttr(
-        axis_gen_op->dyn_cast<paddle::dialect::FullIntArrayOp>(), "value");
-    if (axis.size() == 0) {
-      reduce_all = true;
-    }
-    return details::ReduceInferDim(
-        op, shape_analysis, axis, keepdim, reduce_all);
-  } else {
-    // TODO(lanxianghit): deal with other source: pir::VectorType,
-    // paddle::dialect::DenseTensorType
-    PADDLE_THROW(
-        phi::errors::Unimplemented("SumOpInferSymbolicShape: 'axis' only "
-                                   "support FullIntArrayOp's result now."));
-  }
-
-  return true;
-}
-
-bool ProdOpInferSymbolicShape(pir::Operation *op,
-                              pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  bool keepdim = GetBoolAttr(op, "keep_dim");
-  bool reduce_all = GetBoolAttr(op, "reduce_all");
-
-  auto axis_gen_op = op->operand_source(1).defining_op();
-  if (axis_gen_op->isa<paddle::dialect::FullIntArrayOp>()) {
-    std::vector<int64_t> axis = details::GetVectorAttr(
-        axis_gen_op->dyn_cast<paddle::dialect::FullIntArrayOp>(), "value");
-    return details::ReduceInferDim(
-        op, shape_analysis, axis, keepdim, reduce_all);
-  } else {
-    // TODO(lanxianghit): deal with other source: pir::VectorType,
-    // paddle::dialect::DenseTensorType
-    PADDLE_THROW(
-        phi::errors::Unimplemented("ProdOpInferSymbolicShape: 'axis' only "
-                                   "support FullIntArrayOp's result now."));
-  }
-
-  return true;
-}
-
-bool SliceOpInferSymbolicShape(pir::Operation *op,
-                               pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  pir::Value operand_source = op->operand_source(0);
-  pir::Value operand_starts = op->operand_source(1);
-  pir::Value operand_ends = op->operand_source(2);
-  pir::Value res = op->result(0);
-
-  const symbol::ShapeOrDataDimExprs &operand_shape_or_data =
-      shape_analysis->GetShapeOrDataForValue(operand_source);
-  const symbol::ShapeOrDataDimExprs &starts_shape_data =
-      shape_analysis->GetShapeOrDataForValue(operand_starts);
-  const symbol::ShapeOrDataDimExprs &ends_shape_data =
-      shape_analysis->GetShapeOrDataForValue(operand_ends);
-
-  std::vector<int64_t> axes_vec = details::GetVectorAttr(op, "axes");
-
-  // // Currently, we DO NOT support any element in `starts` is a Symbol.
-  ExprVec starts = slice_utils::GetExprVecFromData(starts_shape_data);
-  ExprVec ends = slice_utils::GetExprVecFromData(ends_shape_data);
-
-  std::vector<int64_t> infer_flags = details::GetVectorAttr(op, "infer_flags");
-
-  const std::vector<int64_t> decrease_axis =
-      details::GetVectorAttr(op, "decrease_axis");
-
-  shape_analysis->SetShapeOrDataForValue(
-      res,
-      slice_utils::SliceRawInferSymbolicShape(operand_shape_or_data,
-                                              starts,
-                                              ends,
-                                              axes_vec,
-                                              infer_flags,
-                                              decrease_axis));
-
-  return true;
-}
-
-bool ConcatOpInferSymbolicShape(
-    pir::Operation *op, pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  pir::Value operand_source = op->operand_source(0);
-  const auto &shape_data_list =
-      shape_analysis->GetShapeOrDataForValue(operand_source)
-          .dyn_cast<symbol::TensorListShapeOrDataDimExprs>();
-
-  CHECK(op->operand_source(1).defining_op()->isa<paddle::dialect::FullOp>());
-
-  int64_t axis = op->operand_source(1)
-                     .defining_op<paddle::dialect::FullOp>()
-                     .attributes()
-                     .at("value")
-                     .dyn_cast<paddle::dialect::ScalarAttribute>()
-                     .data()
-                     .to<int64_t>();
-  size_t rank = shape_data_list[0].shape().size();
-  axis = axis >= 0 ? axis : std::max(int64_t(0), int64_t(axis + rank));
-
-  if (shape_data_list[0].data().has_value()) {
-    if (rank == 1) {
-      ExprVec data = details::GetExprVecFromData(
-          shape_analysis->GetShapeOrDataForValue(operand_source));
-      const std::vector<symbol::DimExpr> shape{std::int64_t(data.size())};
-      symbol::ShapeOrDataDimExprs shape_data{
-          symbol::TensorShapeOrDataDimExprs(shape, data)};
-      pir::Value res = op->result(0);
-      shape_analysis->SetShapeOrDataForValue(res, shape_data);
-
-      return true;
-    } else {
-      PADDLE_THROW(phi::errors::Unimplemented(
-          op->name() +
-          " 's InferSymbolicShape can NOT deal with rank > 1 now."));
-    }
-    std::vector<symbol::DimExpr> data;
-    data.reserve(shape_data_list.size());
-    for (auto &data_elem : shape_data_list) {
-      data.push_back(data_elem.data().value()[0]);
-    }
-    const std::vector<symbol::DimExpr> shape{std::int64_t(data.size())};
-    symbol::ShapeOrDataDimExprs shape_data{
-        symbol::TensorShapeOrDataDimExprs(shape, data)};
-    pir::Value res = op->result(0);
-    shape_analysis->SetShapeOrDataForValue(res, shape_data);
-
-    return true;
-  }
-
-  const std::vector<symbol::DimExpr> &out_dims = [&] {
-    std::vector<symbol::DimExpr> out_dims = shape_data_list[0].shape();
-    for (size_t i = 0; i < rank; ++i) {
-      if (i != static_cast<size_t>(axis)) {
-        details::BuildCstrEqForTensorListAlongAxis(
-            shape_analysis, shape_data_list, i);
-        continue;
-      }
-      for (size_t j = 1; j < shape_data_list.size(); ++j) {
-        out_dims[axis] = out_dims[axis] + shape_data_list[j].shape()[axis];
-      }
-    }
-    return out_dims;
-  }();
-
-  symbol::ShapeOrDataDimExprs shape_data{
-      symbol::TensorShapeOrDataDimExprs(out_dims)};
-
-  pir::Value res = op->result(0);
-  shape_analysis->SetShapeOrDataForValue(res, shape_data);
-
-  return true;
-}
-
-bool GatherNdOpInferSymbolicShape(
-    pir::Operation *op, pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  const auto &x_shape_or_data =
-      shape_analysis->GetShapeOrDataForValue(op->operand_source(0));
-  const auto &index_shape_or_data =
-      shape_analysis->GetShapeOrDataForValue(op->operand_source(1));
-
-  const std::vector<symbol::DimExpr> &x_sym_shape =
-      x_shape_or_data.data().has_value() ? x_shape_or_data.data().value()
-                                         : x_shape_or_data.shape();
-
-  const std::vector<symbol::DimExpr> &index_sym_shape =
-      index_shape_or_data.data().has_value()
-          ? index_shape_or_data.data().value()
-          : index_shape_or_data.shape();
-
-  int x_dims_size = x_sym_shape.size();
-  int index_dims_size = index_sym_shape.size();
-
-  std::vector<symbol::DimExpr> result_sym_dims;
-  // The result dims is
-  //   Index.shape[:-1] + X.shape[Index.shape[-1]:]
-  for (int i = 0; i < index_dims_size - 1; ++i) {
-    result_sym_dims.emplace_back(index_sym_shape[i]);
-  }
-
-  PADDLE_ENFORCE_EQ(
-      index_sym_shape[index_dims_size - 1].Has<std::int64_t>(),
-      true,
-      phi::errors::InvalidArgument(
-          "in GatherNdOpInferSymbolicShape: index[-1] should be unknown"));
-
-  for (int i = static_cast<int>(
-           index_sym_shape[index_dims_size - 1].Get<std::int64_t>());
-       i < x_dims_size;
-       ++i) {
-    result_sym_dims.emplace_back(x_sym_shape[i]);
-  }
-
-  symbol::ShapeOrDataDimExprs shape_data{
-      symbol::TensorShapeOrDataDimExprs(result_sym_dims)};
-
-  pir::Value res = op->result(0);
-  shape_analysis->SetShapeOrDataForValue(res, shape_data);
-
-  return true;
-}
-
-bool TileOpInferSymbolicShape(pir::Operation *op,
-                              pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  pir::Value operand_x = op->operand_source(0);
-  symbol::ShapeOrDataDimExprs x_shape_or_data =
-      shape_analysis->GetShapeOrDataForValue(operand_x);
-  pir::Value operand_repeat_times = op->operand_source(1);
-  symbol::ShapeOrDataDimExprs repeat_times_shape_or_data =
-      shape_analysis->GetShapeOrDataForValue(operand_repeat_times);
-
-  std::vector<symbol::DimExpr> x_dimexpr;
-  if (x_shape_or_data.data().has_value()) {
-    x_dimexpr = x_shape_or_data.data().value();
-  } else {
-    x_dimexpr = x_shape_or_data.shape();
-  }
-
-  std::vector<symbol::DimExpr> repeat_times_dimexpr;
-  if (repeat_times_shape_or_data.data().has_value()) {
-    repeat_times_dimexpr = repeat_times_shape_or_data.data().value();
-  } else {
-    repeat_times_dimexpr = repeat_times_shape_or_data.shape();
-  }
-  if (repeat_times_dimexpr.empty()) {
-    repeat_times_dimexpr = std::vector<symbol::DimExpr>(x_dimexpr.size(), 1);
-  }
-
-  auto out_rank = std::max(static_cast<size_t>(x_dimexpr.size()),
-                           repeat_times_dimexpr.size());
-  std::vector<symbol::DimExpr> out_shape(out_rank);
-  if (x_dimexpr.size() > repeat_times_dimexpr.size()) {
-    auto diff = x_dimexpr.size() - repeat_times_dimexpr.size();
-    repeat_times_dimexpr.insert(repeat_times_dimexpr.begin(), diff, 1);
-  } else {
-    auto diff = repeat_times_dimexpr.size() - x_dimexpr.size();
-    x_dimexpr.insert(x_dimexpr.begin(), diff, 1);
-  }
-
-  for (size_t i = 0; i < repeat_times_dimexpr.size(); ++i) {
-    out_shape[i] = x_dimexpr[i] * repeat_times_dimexpr[i];
-  }
-
-  symbol::ShapeOrDataDimExprs shape_data{
-      symbol::TensorShapeOrDataDimExprs(out_shape)};
-
-  pir::Value res = op->result(0);
-  shape_analysis->SetShapeOrDataForValue(res, shape_data);
-
-  return true;
-}
-
-bool TransposeOpInferSymbolicShape(
-    pir::Operation *op, pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  std::vector<pir::Attribute> perm =
-      op->attributes().at("perm").dyn_cast<pir::ArrayAttribute>().AsVector();
-  if (perm.size() == 1) {
-    // perm must be [0], which means nothing to do with input, just copy the
-    // info from input
-    shape_analysis->SetShapeOrDataForValue(
-        op->result(0),
-        shape_analysis->GetShapeOrDataForValue(op->operand_source(0)));
-    return true;
-  }
-  const std::vector<symbol::DimExpr> &x_dims = [&] {
-    std::vector<symbol::DimExpr> dims;
-    const auto &x_shape_or_data =
-        shape_analysis->GetShapeOrDataForValue(op->operand_source(0));
-    if (x_shape_or_data.data().has_value()) {
-      dims = x_shape_or_data.data().value();
-    } else {
-      dims = x_shape_or_data.shape();
-    }
-    return dims;
-  }();
-
-  int x_rank = x_dims.size();
-
-  const std::vector<int32_t> formatted_axis = [op, x_rank, &perm] {
-    std::vector<int32_t> out(perm.size(), 0);
-    std::transform(perm.begin(),
-                   perm.end(),
-                   out.begin(),
-                   [](pir::Attribute &p) -> int32_t {
-                     return p.dyn_cast<pir::Int32Attribute>().data();
-                   });
-
-    // format the negative axis
-    std::for_each(out.begin(), out.end(), [x_rank](int32_t &v) {
-      if (v < 0) {
-        v += x_rank;
-      }
-    });
-    return out;
-  }();
-
-  int axis_size = static_cast<int>(formatted_axis.size());
-
-  std::vector<symbol::DimExpr> out_dims(x_dims);
-  for (int i = 0; i < axis_size; ++i) {
-    out_dims[i] = x_dims[formatted_axis[i]];
-  }
-
-  shape_analysis->SetShapeOrDataForValue(op->result(0),
-                                         ShapeOrData{TensorExprs(out_dims)});
-
-  return true;
-}
-bool Transpose_OpInferSymbolicShape(
-    pir::Operation *op, pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  return TransposeOpInferSymbolicShape(op, shape_analysis);
-}
-
-bool EmbeddingOpInferSymbolicShape(
-    pir::Operation *op, pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  const auto x_shape_or_data =
-      shape_analysis->GetShapeOrDataForValue(op->operand_source(0));
-  const auto weight_shape_or_data =
-      shape_analysis->GetShapeOrDataForValue(op->operand_source(1));
-  const std::vector<symbol::DimExpr> &x_dims = [&] {
-    std::vector<symbol::DimExpr> dims;
-    if (x_shape_or_data.data().has_value()) {
-      dims = x_shape_or_data.data().value();
-    } else {
-      dims = x_shape_or_data.shape();
-    }
-    return dims;
-  }();
-
-  const std::vector<symbol::DimExpr> &weight_dims = [&] {
-    std::vector<symbol::DimExpr> dims;
-    if (weight_shape_or_data.data().has_value()) {
-      dims = weight_shape_or_data.data().value();
-    } else {
-      dims = weight_shape_or_data.shape();
-    }
-    return dims;
-  }();
-
-  const symbol::ShapeOrDataDimExprs &shape_data = [&] {
-    std::vector<symbol::DimExpr> out_dims = x_dims;
-    // no need to check validation of weight_dims index, since all checks have
-    // been done at corresponding InferMeta
-    out_dims.emplace_back(weight_dims[1]);
-    return symbol::ShapeOrDataDimExprs{
-        symbol::TensorShapeOrDataDimExprs(out_dims)};
-  }();
-
-  shape_analysis->SetShapeOrDataForValue(op->result(0), shape_data);
-
-  return true;
-}
-
-bool SparseWeightEmbeddingOpInferSymbolicShape(
-    pir::Operation *op, pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  PADDLE_THROW(phi::errors::Unimplemented(
-      op->name() + " 's InferSymbolicShape interface is NOT implemented now."));
-  return true;
-}
-
-bool MatmulOpInferSymbolicShape(
-    pir::Operation *op, pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  // x_dims can't be const or ref here, in case to be broadcasted
-  std::vector<symbol::DimExpr> x_dims = [&] {
-    std::vector<symbol::DimExpr> dims;
-    const auto &x_shape_or_data =
-        shape_analysis->GetShapeOrDataForValue(op->operand_source(0));
-    if (x_shape_or_data.data().has_value()) {
-      dims = x_shape_or_data.data().value();
-    } else {
-      dims = x_shape_or_data.shape();
-    }
-    return dims;
-  }();
-
-  // y_dims can't be const or ref here, in case to be broadcasted
-  std::vector<symbol::DimExpr> y_dims = [&] {
-    std::vector<symbol::DimExpr> dims;
-    const auto y_shape_or_data =
-        shape_analysis->GetShapeOrDataForValue(op->operand_source(1));
-    if (y_shape_or_data.data().has_value()) {
-      dims = y_shape_or_data.data().value();
-    } else {
-      dims = y_shape_or_data.shape();
-    }
-    return dims;
-  }();
-
-  size_t ndims_x = x_dims.size();
-  size_t ndims_y = y_dims.size();
-
-  const bool x_broadcasted = [&] {
-    bool broadcasted = false;
-    if (ndims_x == 1) {
-      x_dims.insert(x_dims.begin(), 1);
-      ndims_x = 2;
-      broadcasted = true;
-    }
-    return broadcasted;
-  }();
-
-  const bool y_broadcasted = [&] {
-    bool broadcasted = false;
-    if (ndims_y == 1) {
-      y_dims.emplace_back(1);
-      ndims_y = 2;
-      broadcasted = true;
-    }
-    return broadcasted;
-  }();
-
-  std::vector<symbol::DimExpr> out_dims;
-  if (ndims_x > ndims_y) {
-    out_dims.assign(x_dims.begin(), x_dims.end() - 2);
-  } else if (ndims_x < ndims_y) {
-    out_dims.assign(y_dims.begin(), y_dims.end() - 2);
-  } else {
-    symbol::DimExprBuilder builder{nullptr};
-    for (size_t i = 0; i < ndims_x - 2; ++i) {
-      out_dims.emplace_back(builder.Broadcast(x_dims[i], y_dims[i]));
-    }
-  }
-
-  bool transpose_x_attr = GetBoolAttr(op, "transpose_x");
-  bool transpose_y_attr = GetBoolAttr(op, "transpose_y");
-  symbol::DimExpr out_M =
-      transpose_x_attr ? x_dims[ndims_x - 1] : x_dims[ndims_x - 2];
-  symbol::DimExpr out_N =
-      transpose_y_attr ? y_dims[ndims_y - 2] : y_dims[ndims_y - 1];
-  if (!x_broadcasted) {
-    out_dims.emplace_back(out_M);
-  }
-  if (!y_broadcasted) {
-    out_dims.emplace_back(out_N);
-  }
-
-  shape_analysis->SetShapeOrDataForValue(op->result(0),
-                                         ShapeOrData{TensorExprs(out_dims)});
-
-  if ((ndims_x == ndims_y) && ndims_x >= 2) {
-    if (transpose_x_attr == false && transpose_y_attr == false) {
-      shape_analysis->CreateDimExprBuilder().CstrEq(x_dims[ndims_x - 1],
-                                                    y_dims[ndims_x - 2]);
-    } else if (transpose_x_attr == false && transpose_y_attr == true) {
-      shape_analysis->CreateDimExprBuilder().CstrEq(x_dims[ndims_x - 1],
-                                                    y_dims[ndims_x - 1]);
-    } else if (transpose_x_attr == true && transpose_y_attr == false) {
-      shape_analysis->CreateDimExprBuilder().CstrEq(x_dims[ndims_x - 2],
-                                                    y_dims[ndims_x - 2]);
-    } else {
-      shape_analysis->CreateDimExprBuilder().CstrEq(x_dims[ndims_x - 2],
-                                                    y_dims[ndims_x - 1]);
-    }
-
-    for (size_t i = 0; i < ndims_x - 2; ++i) {
-      shape_analysis->CreateDimExprBuilder().CstrEq(x_dims[i], y_dims[i]);
-    }
-  }
-  return true;
-}
-
-bool MaxOpInferSymbolicShape(pir::Operation *op,
-                             pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  bool keepdim = GetBoolAttr(op, "keepdim");
-
-  const std::vector<int64_t> axis = [&] {
-    pir::Operation *axis_gen_op = op->operand_source(1).defining_op();
-    std::vector<int64_t> axis_vec;
-    if (axis_gen_op->isa<paddle::dialect::FullIntArrayOp>()) {
-      axis_vec = details::GetVectorAttr(
-          axis_gen_op->dyn_cast<paddle::dialect::FullIntArrayOp>(), "value");
-    } else {
-      // TODO(lanxianghit): there's other source: pir::VectorType,
-      // paddle::dialect::DenseTensorType, but after PRIM, maybe always
-      // FullIntArrayOp, to be confirmed
-      PADDLE_THROW(
-          phi::errors::Unimplemented("MaxOpInferSymbolicShape: 'axis' only "
-                                     "support FullIntArrayOp's result now."));
-    }
-    return axis_vec;
-  }();
-
-  bool reduce_all = axis.size() == 0 ? true : false;
-
-  return details::ReduceInferDim(op, shape_analysis, axis, keepdim, reduce_all);
-}
-
-bool WhereOpInferSymbolicShape(pir::Operation *op,
-                               pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  shape_analysis->SetShapeOrDataForValue(
-      op->result(0),
-      shape_analysis->GetShapeOrDataForValue(op->operand_source(0)));
-
-  const std::vector<pir::Value> &operands = {op->operand_source(0),
-                                             op->operand_source(1)};
-
-  size_t rank = shape_analysis->GetShapeOrDataForValue(op->operand_source(0))
-                    .shape()
-                    .size();
-
-  for (size_t i = 0; i < rank; ++i) {
-    paddle::dialect::details::BuildCstrEqForTensorListAlongAxis(
-        shape_analysis, operands, i);
-  }
-
-  return true;
-}
-
-bool Where_OpInferSymbolicShape(
-    pir::Operation *op, pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  return WhereOpInferSymbolicShape(op, shape_analysis);
-}
-
-bool FeedOpInferSymbolicShape(pir::Operation *op,
-                              pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  const common::DDim &result_dims =
-      op->result(0).type().dyn_cast<pir::DenseTensorType>().dims();
-  std::vector<symbol::DimExpr> out_dims;
-  for (int i = 0; i < result_dims.size(); i++) {
-    if (result_dims[i] == -1) {
-      out_dims.emplace_back(shape_analysis->GetNextSymName());
-    } else {
-      out_dims.emplace_back(result_dims[i]);
-    }
-  }
-
-  shape_analysis->SetShapeOrDataForValue(
-      op->result(0),
-      symbol::ShapeOrDataDimExprs{symbol::TensorShapeOrDataDimExprs(out_dims)});
-
-  return true;
-}
-
-bool TopPSamplingOpInferSymbolicShape(
-    pir::Operation *op, pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  const auto &x_dims = [op, shape_analysis] {
-    const auto &shape_or_data =
-        shape_analysis->GetShapeOrDataForValue(op->operand_source(0));
-    if (shape_or_data.data().has_value()) {
-      return shape_or_data.data().value();
-    } else {
-      return shape_or_data.shape();
-    }
-  }();
-
-  // all the result have the same shape
-  for (uint32_t rst_idx = 0; rst_idx < op->num_results(); rst_idx++) {
-    const std::vector<symbol::DimExpr> out_dims{x_dims[0], 1};
-    shape_analysis->SetShapeOrDataForValue(
-        op->result(rst_idx),
-        symbol::ShapeOrDataDimExprs{
-            symbol::TensorShapeOrDataDimExprs(out_dims)});
-  }
-
-  return true;
-}
-
-bool ExpandAsOpInferSymbolicShape(
-    pir::Operation *op, pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  PADDLE_THROW(phi::errors::Unimplemented(
-      op->name() + " 's InferSymbolicShape interface is NOT implemented now."));
-  return true;
-}
-
-bool SplitOpInferSymbolicShape(pir::Operation *op,
-                               pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  // input
-  const auto &x_shape_or_data =
-      shape_analysis->GetShapeOrDataForValue(op->operand_source(0));
-  PADDLE_ENFORCE_EQ(x_shape_or_data.data().has_value(),
-                    false,
-                    phi::errors::InvalidArgument(
-                        "InferSymbolicShape of SplitOp only support input with "
-                        "value now."));
-  const auto &x_dims_sym = x_shape_or_data.shape();
-
-  // axis
-  CHECK(op->operand_source(2).defining_op()->isa<paddle::dialect::FullOp>());
-
-  int64_t axis = op->operand_source(2)
-                     .defining_op<paddle::dialect::FullOp>()
-                     .attributes()
-                     .at("value")
-                     .dyn_cast<paddle::dialect::ScalarAttribute>()
-                     .data()
-                     .to<int64_t>();
-
-  // sections
-  const std::vector<symbol::DimExpr> &sections_sym = [&] {
-    const auto &sections_shape_or_data =
-        shape_analysis->GetShapeOrDataForValue(op->operand_source(1));
-    std::vector<symbol::DimExpr> sections_sym;
-    if (sections_shape_or_data.data().has_value()) {
-      sections_sym = sections_shape_or_data.data().value();
-    } else {
-      sections_sym = sections_shape_or_data.shape();
-    }
-    return sections_sym;
-  }();
-
-  // output
-  const symbol::TensorListShapeOrDataDimExprs &output_shape_data_list = [&] {
-    const auto &GetSum = [&](const auto &dim_exprs, const auto &Filter) {
-      symbol::DimExpr sum{0};
-      for (const auto &dim_expr : dim_exprs) {
-        if (Filter(dim_expr)) {
-          sum = sum + dim_expr;
-        }
-      }
-      return sum;
-    };
-    const auto &All = [&](const auto &dim_exprs, const auto &Cond) {
-      for (const auto &dim_expr : dim_exprs) {
-        if (!Cond(dim_expr)) {
-          return false;
-        }
-      }
-      return true;
-    };
-    const auto &IsNotMinusOne = [&](const symbol::DimExpr &dim_expr) {
-      if (dim_expr.isa<int64_t>()) {
-        return dim_expr.dyn_cast<int64_t>() != static_cast<int64_t>(-1);
-      }
-      return true;
-    };
-    const auto &sum_exclude_minus_one = GetSum(sections_sym, IsNotMinusOne);
-
-    const bool &all_sections_sym_not_minus_one =
-        All(sections_sym, IsNotMinusOne);
-    if (all_sections_sym_not_minus_one) {
-      shape_analysis->CreateDimExprBuilder().CstrEq(x_dims_sym[axis],
-                                                    sum_exclude_minus_one);
-    }
-
-    symbol::TensorListShapeOrDataDimExprs shape_data_list;
-    std::vector<symbol::DimExpr> output_dims_sym = x_dims_sym;
-    if (!all_sections_sym_not_minus_one && sections_sym.size() == 1) {
-      VLOG(3) << "[SplitOp]-1 is the only split section. The output shape is "
-                 "identical to the input shape.";
-      shape_data_list.push_back(
-          symbol::TensorShapeOrDataDimExprs(output_dims_sym));
-      return shape_data_list;
-    }
-    for (uint32_t idx = 0; idx < sections_sym.size(); idx++) {
-      const auto &section_sym = sections_sym[idx];
-      output_dims_sym[axis] = IsNotMinusOne(section_sym)
-                                  ? section_sym
-                                  : x_dims_sym[axis] - sum_exclude_minus_one;
-
-      shape_data_list.push_back(
-          symbol::TensorShapeOrDataDimExprs(output_dims_sym));
-    }
-    return shape_data_list;
-  }();
-
-  shape_analysis->SetShapeOrDataForValue(
-      op->result(0), symbol::ShapeOrDataDimExprs{output_shape_data_list});
-
-  return true;
-}
-
-//  Not Implemented Ops.
-bool GatherOpInferSymbolicShape(
-    pir::Operation *op, pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  const auto &input_shape_or_data =
-      shape_analysis->GetShapeOrDataForValue(op->operand_source(0));
-  const auto &index_shape_or_data =
-      shape_analysis->GetShapeOrDataForValue(op->operand_source(1));
-
-  const auto &numel = [&] {
-    symbol::DimExpr numel{1};
-    for (const auto &dim_expr : index_shape_or_data.shape()) {
-      numel = numel * dim_expr;
-    }
-    return numel;
-  }();
-
-  const auto &axis_shape_or_data =
-      shape_analysis->GetShapeOrDataForValue(op->operand_source(2));
-
-  const std::vector<symbol::DimExpr> &input_sym_shape =
-      input_shape_or_data.data().has_value()
-          ? input_shape_or_data.data().value()
-          : input_shape_or_data.shape();
-
-  const std::vector<symbol::DimExpr> &index_sym_shape =
-      index_shape_or_data.data().has_value()
-          ? index_shape_or_data.data().value()
-          : index_shape_or_data.shape();
-
-  int axis =
-      static_cast<int>(axis_shape_or_data.data().value()[0].Get<int64_t>());
-  if (axis < 0) axis += input_sym_shape.size();
-
-  const auto &out_sym_shape = [&] {
-    std::vector<symbol::DimExpr> out_sym_shape;
-
-    if (index_sym_shape.size() == 0) {
-      if (input_sym_shape.size() == 1) {
-        out_sym_shape.push_back(symbol::DimExpr{0});
-      } else {
-        for (int i = 0; i < axis; ++i) {
-          out_sym_shape.push_back(input_sym_shape[i]);
-        }
-        for (size_t i = axis + 1; i < input_sym_shape.size(); ++i) {
-          out_sym_shape.push_back(input_sym_shape[i]);
-        }
-      }
-    } else {
-      for (int i = 0; i < axis; ++i) {
-        out_sym_shape.push_back(input_sym_shape[i]);
-      }
-      out_sym_shape.push_back(numel);
-      for (size_t i = axis + 1; i < input_sym_shape.size(); ++i) {
-        out_sym_shape.push_back(input_sym_shape[i]);
-      }
-    }
-    return out_sym_shape;
-  }();
-
-  symbol::ShapeOrDataDimExprs shape_data{
-      symbol::TensorShapeOrDataDimExprs(out_sym_shape)};
-
-  pir::Value res = op->result(0);
-  shape_analysis->SetShapeOrDataForValue(res, shape_data);
-
-  return true;
-}
-
-bool KronOpInferSymbolicShape(pir::Operation *op,
-                              pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  const auto &x_shape_or_data =
-      shape_analysis->GetShapeOrDataForValue(op->operand_source(0)).shape();
-  const auto &y_shape_or_data =
-      shape_analysis->GetShapeOrDataForValue(op->operand_source(1)).shape();
-  const int rank_x = x_shape_or_data.size();
-  const int rank_y = y_shape_or_data.size();
-  const int rank = (rank_x > rank_y) ? rank_x : rank_y;
-
-  std::vector<symbol::DimExpr> dim_out;
-  dim_out.reserve(rank);
-  const auto one = symbol::DimExpr{1};
-  const auto minus_one = symbol::DimExpr{-1};
-  for (int i = 0; i < rank; i++) {
-    symbol::DimExpr dim_xi =
-        (i < rank - rank_x) ? one : x_shape_or_data.at(i - (rank - rank_x));
-    symbol::DimExpr dim_yi =
-        (i < rank - rank_y) ? one : y_shape_or_data.at(i - (rank - rank_y));
-    dim_out.push_back(dim_xi * dim_yi);
-  }
-  symbol::ShapeOrDataDimExprs shape_data{
-      symbol::TensorShapeOrDataDimExprs(dim_out)};
-  pir::Value res = op->result(0);
-  shape_analysis->SetShapeOrDataForValue(res, shape_data);
-  return true;
-}
-
-//  Not Impelmented Ops.
-bool LogcumsumexpOpInferSymbolicShape(
-    pir::Operation *op, pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  PADDLE_THROW(phi::errors::Unimplemented(
-      op->name() + " 's InferSymbolicShape interface is NOT implemented now."));
-  return true;
-}
-
-bool MaskedSelectOpInferSymbolicShape(
-    pir::Operation *op, pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  PADDLE_THROW(phi::errors::Unimplemented(
-      op->name() + " 's InferSymbolicShape interface is NOT implemented now."));
-  return true;
-}
-bool PoissonOpInferSymbolicShape(
-    pir::Operation *op, pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  PADDLE_THROW(phi::errors::Unimplemented(
-      op->name() + " 's InferSymbolicShape interface is NOT implemented now."));
-  return true;
-}
-
-bool SearchsortedOpInferSymbolicShape(
-    pir::Operation *op, pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  PADDLE_THROW(phi::errors::Unimplemented(
-      op->name() + " 's InferSymbolicShape interface is NOT implemented now."));
-  return true;
-}
-
-bool TakeAlongAxisOpInferSymbolicShape(
-    pir::Operation *op, pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  // input
-  const auto &arr_shape_or_data =
-      shape_analysis->GetShapeOrDataForValue(op->operand_source(0));
-  const auto &indices_shape_or_data =
-      shape_analysis->GetShapeOrDataForValue(op->operand_source(1));
-  const auto &attributes = op->attributes();
-  int axis = attributes.at("axis").dyn_cast<pir::Int32Attribute>().data();
-
-  const std::vector<symbol::DimExpr> &arr_sym_shape =
-      arr_shape_or_data.data().has_value() ? arr_shape_or_data.data().value()
-                                           : arr_shape_or_data.shape();
-  const std::vector<symbol::DimExpr> &indices_sym_shape =
-      indices_shape_or_data.data().has_value()
-          ? indices_shape_or_data.data().value()
-          : indices_shape_or_data.shape();
-
-  if (axis < 0) axis += arr_sym_shape.size();
-
-  const auto &out_sym_shape = [&] {
-    std::vector<symbol::DimExpr> out_sym_shape;
-    for (int i = 0; i < axis; ++i) {
-      out_sym_shape.push_back(arr_sym_shape[i]);
-    }
-    out_sym_shape.push_back(indices_sym_shape[axis]);
-    for (size_t i = axis + 1; i < arr_sym_shape.size(); ++i) {
-      out_sym_shape.push_back(arr_sym_shape[i]);
-    }
-    return out_sym_shape;
-  }();
-
-  symbol::ShapeOrDataDimExprs shape_data{
-      symbol::TensorShapeOrDataDimExprs(out_sym_shape)};
-
-  pir::Value res = op->result(0);
-  shape_analysis->SetShapeOrDataForValue(res, shape_data);
-
-  return true;
-}
-
-bool TopkOpInferSymbolicShape(pir::Operation *op,
-                              pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  PADDLE_THROW(phi::errors::Unimplemented(
-      op->name() + " 's InferSymbolicShape interface is NOT implemented now."));
-  return true;
-}
-bool UnbindOpInferSymbolicShape(
-    pir::Operation *op, pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  PADDLE_THROW(phi::errors::Unimplemented(
-      op->name() + " 's InferSymbolicShape interface is NOT implemented now."));
-  return true;
-}
-bool UniqueConsecutiveOpInferSymbolicShape(
-    pir::Operation *op, pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  PADDLE_THROW(phi::errors::Unimplemented(
-      op->name() + " 's InferSymbolicShape interface is NOT implemented now."));
-  return true;
-}
-bool LinspaceOpInferSymbolicShape(
-    pir::Operation *op, pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  PADDLE_THROW(phi::errors::Unimplemented(
-      op->name() + " 's InferSymbolicShape interface is NOT implemented now."));
-  return true;
-}
-bool LogspaceOpInferSymbolicShape(
-    pir::Operation *op, pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  PADDLE_THROW(phi::errors::Unimplemented(
-      op->name() + " 's InferSymbolicShape interface is NOT implemented now."));
-  return true;
-}
-bool LogsumexpOpInferSymbolicShape(
-    pir::Operation *op, pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  PADDLE_THROW(phi::errors::Unimplemented(
-      op->name() + " 's InferSymbolicShape interface is NOT implemented now."));
-  return true;
-}
-
-bool MinOpInferSymbolicShape(pir::Operation *op,
-                             pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  PADDLE_THROW(phi::errors::Unimplemented(
-      op->name() + " 's InferSymbolicShape interface is NOT implemented now."));
-  return true;
-}
-
-bool PadOpInferSymbolicShape(pir::Operation *op,
-                             pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  PADDLE_THROW(phi::errors::Unimplemented(
-      op->name() + " 's InferSymbolicShape interface is NOT implemented now."));
-  return true;
-}
-bool RandintOpInferSymbolicShape(
-    pir::Operation *op, pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  PADDLE_THROW(phi::errors::Unimplemented(
-      op->name() + " 's InferSymbolicShape interface is NOT implemented now."));
-  return true;
-}
-
-bool RepeatInterleaveOpInferSymbolicShape(
-    pir::Operation *op, pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  PADDLE_THROW(phi::errors::Unimplemented(
-      op->name() + " 's InferSymbolicShape interface is NOT implemented now."));
-  return true;
-}
-bool SplitWithNumOpInferSymbolicShape(
-    pir::Operation *op, pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  PADDLE_THROW(phi::errors::Unimplemented(
-      op->name() + " 's InferSymbolicShape interface is NOT implemented now."));
-  return true;
-}
-bool TrilIndicesOpInferSymbolicShape(
-    pir::Operation *op, pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  PADDLE_THROW(phi::errors::Unimplemented(
-      op->name() + " 's InferSymbolicShape interface is NOT implemented now."));
-  return true;
-}
-bool TriuIndicesOpInferSymbolicShape(
-    pir::Operation *op, pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  PADDLE_THROW(phi::errors::Unimplemented(
-      op->name() + " 's InferSymbolicShape interface is NOT implemented now."));
-  return true;
-}
-bool UniformOpInferSymbolicShape(
-    pir::Operation *op, pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  PADDLE_THROW(phi::errors::Unimplemented(
-      op->name() + " 's InferSymbolicShape interface is NOT implemented now."));
-  return true;
-}
-bool UniqueOpInferSymbolicShape(
-    pir::Operation *op, pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  PADDLE_THROW(phi::errors::Unimplemented(
-      op->name() + " 's InferSymbolicShape interface is NOT implemented now."));
-  return true;
-}
-
-bool FullWithTensorOpInferSymbolicShape(
-    pir::Operation *op, pir::ShapeConstraintIRAnalysis *shape_analysis) {
-  pir::Value operand_source = op->operand_source(0);
-  const symbol::ShapeOrDataDimExprs &operand_shape_or_data =
-      shape_analysis->GetShapeOrDataForValue(operand_source);
-
-  const auto &out_shape = operand_shape_or_data.data().has_value()
-                              ? operand_shape_or_data.data().value()
-                              : operand_shape_or_data.shape();
-
-  shape_analysis->SetShapeOrDataForValue(
-      op->result(0), symbol::TensorShapeOrDataDimExprs(out_shape));
-  return true;
-}
-}  // namespace paddle::dialect
