--- paddle/fluid/pir/dialect/operator/interface/infer_symbolic_shape/paddle_op_infer_sym.h
@@ -1,69 +1,0 @@
-// Copyright (c) 2023 PaddlePaddle Authors. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#pragma once
-
-#include "paddle/pir/include/dialect/shape/utils/shape_analysis.h"
-
-namespace paddle::dialect {
-
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Shape)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(ShapeSr)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Stack)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Sum)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Slice)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Concat)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(GatherNd)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Tile)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Transpose)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Transpose_)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Prod)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(TakeAlongAxis)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Embedding)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(SparseWeightEmbedding)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Matmul)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Max)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Where)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Where_)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Feed)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(TopPSampling)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(ExpandAs)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Split)
-
-//  Not Impelmented Ops.
-
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Gather)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Kron)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Logcumsumexp)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(MaskedSelect)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Poisson)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Searchsorted)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Topk)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Unbind)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(UniqueConsecutive)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Linspace)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Logspace)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Logsumexp)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Min)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Pad)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Randint)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(RepeatInterleave)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(SplitWithNum)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(TrilIndices)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(TriuIndices)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Uniform)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Unique)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(FullWithTensor)
-
-}  // namespace paddle::dialect
