+++ paddle/fluid/pir/dialect/operator/interface/infer_symbolic_shape/same_operands_result.h
@@ -107,6 +107,8 @@ OP_DECLARE_INFER_SYMBOLIC_SHAPE(Sin)
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(Sin_)
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(Sinh)
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(Sinh_)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(Softmax)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(Softmax_)
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(Tan)
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(Tan_)
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(Tanh)
@@ -115,8 +117,6 @@ OP_DECLARE_INFER_SYMBOLIC_SHAPE(Tril)
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(Tril_)
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(Trunc)
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(Trunc_)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Softmax)
-OP_DECLARE_INFER_SYMBOLIC_SHAPE(Softmax_)
 
 }  // namespace paddle::dialect
