+++ paddle/fluid/pir/dialect/operator/interface/infer_symbolic_shape/unary_infer_sym.h
@@ -17,7 +17,6 @@
 #include "paddle/pir/include/dialect/shape/utils/shape_analysis.h"
 
 namespace paddle::dialect {
-
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(Argmax)
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(Argmin)
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(AsComplex)
@@ -32,10 +31,31 @@ OP_DECLARE_INFER_SYMBOLIC_SHAPE(DiagEmbed)
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(Diagonal)
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(Einsum)
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(Kthvalue)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(Logcumsumexp)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(Logsumexp)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(Max)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(Min)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(Pad)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(Poisson)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(Prod)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(RepeatInterleave)
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(Reshape)
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(Reshape_)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(Shape)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(ShapeSr)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(Slice)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(Split)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(SplitWithNum)
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(Squeeze)
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(Squeeze_)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(Sum)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(Tile)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(Topk)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(Transpose)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(Transpose_)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(Unbind)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(Unique)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(UniqueConsecutive)
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(Unsqueeze)
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(Unsqueeze_)
